Sort chunks before adding to queue
An additional advantage is that the last host to pick a chunk will
probably receive one that is smaller than the others (since the last
chunk is most likely shorter), which means it might not even finish
last just because it was the last to start. So we have more time spent
with all hosts working in parallel, which makes the whole process take
less time.