Move to zero-capacity channel for encoder queue
Previously the host thread would have one chunk in the queue and one
ready to put into the queue as soon as the encoder caught up, which
resulted in two idle chunks lying around. One is enough and could in
fact lead to better performance, since with fewer chunks "reserved" this
way, the more powerful hosts can do more of the work and have less idle
time.